[ios] Move keyboard and keyboardview to ios-common
common keyboard related code moved to ios-common
