+++ xbmc/platform/darwin/ios-common/CMakeLists.txt
@@ -1,5 +1,9 @@
-set(SOURCES AnnounceReceiver.mm)
+set(SOURCES AnnounceReceiver.mm
+            IOSKeyboard.mm
+            IOSKeyboardView.mm)
 
-set(HEADERS AnnounceReceiver.h)
+set(HEADERS AnnounceReceiver.h
+            IOSKeyboard.h
+            IOSKeyboardView.h)
 
 core_add_library(platform_ios-common)
