+++ xbmc/platform/darwin/ios-common/IOSKeyboard.h
@@ -13,11 +13,11 @@
 class CIOSKeyboard : public CGUIKeyboard
 {
   public:
-    CIOSKeyboard():m_pCharCallback(NULL),m_bCanceled(false){}
+    CIOSKeyboard():m_pCharCallback(nullptr),m_bCanceled(false){}
     virtual bool ShowAndGetInput(char_callback_t pCallback, const std::string &initialString, std::string &typedString, const std::string &heading, bool bHiddenInput);
     virtual void Cancel();
     void fireCallback(const std::string &str);
-    void invalidateCallback() {m_pCharCallback = NULL;}
+    void invalidateCallback() {m_pCharCallback = nullptr;}
     virtual bool SetTextToKeyboard(const std::string &text, bool closeKeyboard = false);
 
   private:
