+++ xbmc/platform/darwin/ios/CMakeLists.txt
@@ -1,14 +1,10 @@
 set(SOURCES IOSEAGLView.mm
             IOSExternalTouchController.mm
-            IOSKeyboard.mm
-            IOSKeyboardView.mm
             IOSScreenManager.mm
             XBMCController.mm)
 
 set(HEADERS IOSEAGLView.h
             IOSExternalTouchController.h
-            IOSKeyboard.h
-            IOSKeyboardView.h
             IOSScreenManager.h
             XBMCApplication.h
             XBMCController.h
